requirements.txt
Flask==2.2.5
python-dotenv
argparse
chromadb
langchain
langchain-community
ibm-watson-machine-learning